up to date

+++ kernel/src/cpu.rs
@@ -2,14 +2,12 @@ use core::arch::x86_64::_rdtsc;
 use raw_cpuid::{CpuId, CpuIdReaderNative};
 
 pub fn get_cycles() -> u64 {
-    unsafe {
-        _rdtsc()
-    }
+    unsafe { _rdtsc() }
 }
-pub fn wait_cycle(cycles: u64) {
-    let start = get_cycles();
+pub fn wait_cycle(cycles: u128) {
+    let start = get_cycles() as u128;
     loop {
-        if (get_cycles() >= cycles + start) {
+        if (get_cycles() as u128 >= cycles + start) {
             return;
         }
     }
@@ -19,4 +17,3 @@ pub fn get_cpu_info() -> CpuId<CpuIdReaderNative> {
     let info = CpuId::new();
     info
 }
-
